requirements.txt
streamlit
spotipy
python-dotenv
pandas
scikit-learn
pickle-mixin
